Prevent sensors from disappearing overnight
As the sun goes down, my solar ECU-3 returns empty values for watts etc for each panel as it goes offline. This PR replaces the empty values with "0" (for numeric fields) and a reasonable substitute timestamp. This allows the sensor to stay visible in Home Assistant overnight.

I.E.
`"403000056327-B": ["", "", "", "", ""]`
becomes
`"403000056327-B": ["0", "0", "0", "0", "2023-10-05 18:34:33"]`